+++ .storybook/preview.ts
@@ -7,7 +7,6 @@ import isChromatic from "./isChromatic";
 import sageStorybookTheme from "./sage-storybook-theme";
 import withGlobalStyles from "./with-global-styles";
 import withLocaleSelector from "./with-locale-selector";
-import withPortalProvider from "./with-portal-provider";
 import { withThemeProvider, globalThemeProvider } from "./withThemeProvider";
 import withReducedMotion from "./with-reduced-motion";
 import withFusionTokens from "./with-fusion-tokens";
@@ -92,7 +91,6 @@ const decorators = [
   withThemeProvider,
   withLocaleSelector,
   withFusionTokens,
-  withPortalProvider,
   withReducedMotion,
 ];