+++ src/components/menu/__internal__/submenu/submenu.context.tsx
@@ -10,6 +10,7 @@ export interface SubmenuContextProps {
   ) => void;
   blockIndex?: number;
   submenuHasMaxWidth?: boolean;
+  closeSubmenu?: () => void;
 }
 
 const SubmenuContext = React.createContext<SubmenuContextProps>({});